+++ R/nodes.R
@@ -1,0 +1,6 @@
+#' Nodes
+#' 
+#' @export
+gt_nodes <- function(){
+  
+}