+++ test.R
@@ -1,0 +1,5 @@
+library(rtweet)
+
+twitter_token <- readRDS("twitter_token.rds")
+
+tweets <- search_tweets("#rstats")