+++ NEWS.md
@@ -1,3 +1,9 @@
+# graphTweets 0.4.0
+
+Major release: overhaul to 1) make computations much faster, 2) make the whole package more tidyverse friendly and 3) switch to `rtweet` as main source.
+
+* `getEdges` & `getNodes` are now deprecated in favour of `gt_edges` and `gt_nodes`
+
 # graphTweets 0.3.2
 
 * documentation corrected